requirements.txt
requests>=2.25.0
pyyaml>=5.4.0
matplotlib>=3.3.0
numpy>=1.19.0

## Bioinformatics Tools
biopython>=1.79
pysam>=0.16.0

## Pipeline Specific
multiqc>=1.11
deeptools>=3.5.0

## Development Tools
pytest>=6.0.0
pytest-cov>=2.10.0
